Remove _aqua.py, match numpy version to qiskit terra
