+++ qiskit/aqua/__init__.py
@@ -42,8 +42,9 @@ from .utils.mcu3 import mcu3
 from .quantum_instance import QuantumInstance
 from .operator import Operator
 from .algorithms import QuantumAlgorithm
-from .qiskit_aqua import QiskitAqua
-from ._aqua import run_algorithm, run_algorithm_to_json
+from .qiskit_aqua import (QiskitAqua,
+                          run_algorithm,
+                          run_algorithm_to_json)
 from ._logging import (get_logging_level,
                        build_logging_config,
                        set_logging_config,
