+++ requirements.txt
@@ -1,7 +1,7 @@
 qiskit-terra>=0.8.0,<0.9
 scipy>=0.19,!=0.19.1
 sympy>=1.3
-numpy>=1.13
+numpy>=1.13,<1.16
 psutil>=5
 jsonschema>=2.6,<2.7
 scikit-learn==0.20.0
